refactor: server, client & shared
